Add postprocessing for name parameter of opendir() and __opendir2()
This postprocessing is needed to add mapping between directory name
and directory descriptor to fdpathdb. This is necessary in common
scenario when directory is listed with readdir(), and files inside it
are being open with openat().

The patch has been submited upstream as well, so hopefully it won't be
needed anymore upon the new version release.

Original patch by: Oleg Girko <ol+mer@infoserver.ru>

Signed-off-by: Juha Kallioinen <juha.kallioinen@jollamobile.com>