[modes] Deal with libsolv updates. Contribute to JB#36070
libsolv version must be equal between target and tools root when
enabling acceleration for programs from libsolv-tools, zypper, and –
what is not obvious – also for rpm.  Even if the dependency between rpm
and libsolv is of the opposite direction, it is necessary that both are
using the same db4 build or the result produced by one wouldn't be
usable by the other one.

I decided to simplify it by applying one rule to all package management
related programs.